fix(month-report): 도넛 라벨 좌표의 cos·sin 오버로드를 명시한다
Xcode 26.3은 Double 각도와 CGFloat 반지름 사이에서 cos·sin 오버로드를
결정하지 못해 ambiguous use로 컴파일이 깨진다. 로컬 26.6은 해소해서
CI에서만 드러났다. 결과 좌표는 그대로다.

+++ woni_app/Features/MonthReport/Components/DonutChartView.swift
@@ -64,8 +64,8 @@ struct DonutChartView: View {
     private func labelPosition(for slice: ReportDonutSlice) -> CGPoint {
         let angle = slice.midAngleFraction * 2 * Double.pi - Double.pi / 2
         return CGPoint(
-            x: chartCenter.x + cos(angle) * labelRadius,
-            y: chartCenter.y + sin(angle) * labelRadius
+            x: chartCenter.x + CGFloat(cos(angle)) * labelRadius,
+            y: chartCenter.y + CGFloat(sin(angle)) * labelRadius
         )
     }
 }
